chore: [DHIS2-11393] update PR comments
