+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2021-07-20T06:39:58.471Z\n"
-"PO-Revision-Date: 2021-07-20T06:39:58.471Z\n"
+"POT-Creation-Date: 2021-07-20T08:34:33.116Z\n"
+"PO-Revision-Date: 2021-07-20T08:34:33.116Z\n"
 
 msgid "Choose one or more dates..."
 msgstr ""
@@ -613,9 +613,6 @@ msgstr ""
 msgid "Enrollment{{escape}} New Event"
 msgstr ""
 
-msgid "Save new"
-msgstr ""
-
 msgid "Enrollment{{escape}} View Event"
 msgstr ""
 
@@ -722,6 +719,9 @@ msgid "Create for"
 msgstr ""
 
 msgid "You can also choose a program from the top bar and create in that program"
+msgstr ""
+
+msgid "Save new"
 msgstr ""
 
 msgid "Save new {{trackedEntityTypeName}} and link"
