+++ src/core_modules/capture-core/components/Pages/EnrollmentAddEvent/EnrollmentAddEventPage.types.js
@@ -3,6 +3,6 @@ import { ProgramStage } from '../../../metaData';
 
 export type Props = {|
     programStage: ProgramStage,
-    currentScopeId: string,
+    programId: string,
     ...CssClasses,
 |};
