+++ src/core_modules/capture-core/components/WidgetAddEvent/WidgetAddEvent.types.js
@@ -1,0 +1,10 @@
+// @flow
+
+import { ProgramStage } from '../../metaData';
+
+export type Props = {|
+    programStage: ProgramStage,
+    currentScopeId: string,
+    resultsPageSize: number,
+    ...CssClasses,
+|};
