+++ src/core_modules/capture-core/components/WidgetAddEvent/index.js
@@ -1,0 +1,1 @@
+export { WidgetAddEvent } from './WidgetAddEvent.component';
